chore(user): move to 'core' dir
